# model/model.go
/*
 * AUdit Trail
 *
 * This is a sample server Petstore server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters.
 *
 * API version: 1.0.0
 * Contact: apiteam@swagger.io
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package model

type Event struct {
	// Name of the Event Type. No classification (codification) is assumed, the type may be different depending on the system (numbers, codes, names, etc.)
	EventType string `json:"event-type" validate:"required"`
	// Indicates whether the Event was generated automatically by the system, or it was caused by a user action
	SystemEvent bool `json:"system-event"`
	// Indicates the System where the Event was originated from. It could also be an Id of the system from which the event originated from. If it was generated from a sensor, this field can be the sensor ID. e.g. C232E-1235-EF231-EF3247
	OriginSystem string `json:"origin-system" validate:"required"`
	// Indicates the time at which the Event was created
	CreatedTime string `json:"created-time" validate:"required"`
	// Indicates the time at which the Event was added to the Audit Trail
	AddedTime string `json:"added-time"`
	// Indicates whether the Event was a success or a failure
	Success bool `json:"success"`
	// Indicates the resource or component to which the event is associated with. For example, when a user with Id 9098 is created. The affected-resource could be 9098. If it's a general event, for example a System Failure, it could indicate the id of the System, e.g. the CRM.
	AffectedResource string `json:"affected-resource"`
	// Indicates the IP of the system from where the event was originated
	OriginIP string `json:"origin-IP"`
	// Includes additional data, which is relevant to the type of event, in JSON format
	DynamicData string `json:"dynamic-data"`
	//Indicates the gravity or importance of the Event. No codification was made on this, because the gravity or importance classification may depend on the type of system, and can be different
	Gravity string `json:"gravity"`
}